minstrel_ht-api: rate-setting fill unused txpower
The 'rates' command doesn't require to specify the tx power idx for all
mrr stages, thus less can be specified in a command. For rates, having
the ratetbl zeroed initially and then only setting the specified slots
is sufficient. However, txpower uses a value of -1 to denote 'default'
or 'unspecified'.
This fills the remaining unused txpower slots of the ratetbl with -1.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>