ath9k: fix tx power reporting in txs
Based on the chip revision, the ath9k driver applies some factors and
offsets to the tx power index that is given, but the hardware does not
provide a way report back the used tx power. Thus, only the adjusted tx
power was reported, which significantly differs from the one specified.

This fixes the issue by removing the applied offsets only for reporting,
giving back the corresponding tx power index that is valid for the tx
power range that is currently used for all ath9k chips (0..63 =>
0..31.5 dBm).

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>