ath9k: fix tpc power limiting
ath9k maintains tables the configured maximum power level for each tx
rate. This is mostly to apply user power limits or regulatory limits.
The index of a rate is used as an index into the power table to retrieve
the maximum power level. This table is initialised starting at index 0
with CCK/OFDM rates. However, if an MCS rate is encountered, the index
is not shifted and thus, an incorrect maximum power level is used.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>