minstrel_ht-api: add preferred name for mrr setting command
Until now, the command was named 'rates' because it currently only
allows to set rates and try counts of the multi-rate-retry chain.
In preparation for merging TPC support, the command will be renamed to
'mrr' instead of 'rates' to be more accurate as soon as TX power can
also be set.

However, 'rates' will be kept as alternative name for backward
compatibility so other tools aren't broken immediately.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>