minstrel_ht_api: include API version in api_info
Prints the current version of the Minstrel-HT API as first line in
api_info with the format 'version;<version number>'. This should be used
by clients to deal with major changes between API versions.

The current API version is 1. This should be increased in case of major
changes regarding syntax, etc.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>