minstrel-ht-api: export per-phy info via API
The API creates a new file 'api_phy' in the already used debugfs
directory. This file contains per-phy information, currently the name of
the loaded device driver, a list of the interfaces and the current
operation mode of Minstrel-HT. This bundles required information and
allows arbitrary clients to use it instead of fetching the information
on their own, e.g. through several debugfs calls or directory path
parsing.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>