ath9k: fix txpwr_table debugfs patch
Due to the recently introduced usage of Werror as default for kernel
compilation, the patch caused a frame-length compilation error because
it used a 1024-byte stack-allocated buffer along other variables and the
overall maximum is limited to 1024 bytes.

The size of the buffer is reduced, a lower size is sufficient for the
output anyway.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>