minstrel_ht-api: manual Minstrel mode per STA
This introduces the ability to select between auto and manual Minstrel
mode per STA, previously it was only possible per PHY. This allows to
run user-space RC for one station and kernel-space Minstrel for another
one.

To achieve this, some of the code belonging to 'api_control' needed to
be reorganized execute more commands per STA and still keep the echoing
working and also easily usable for other commands in the future. This
covers the separation of the command processing code true to the motto
'divide and conquer', i.e. separating the minstrel_ht_control_write
function into further processing steps. Furthermore,
minstrel_ht_api_sta_cmd was extended and now also covers the
'reset_stats' command which previously was handled separately although
it is also an STA command.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>