minstrel_ht: properly handle tx power
For now, the txpower_idx field is always be filled with the default
value '-1'. This value means something like 'invalid', 'default' or
'pass to driver'. The driver will decide which transmit power to use. As
long as there is nothing implemented in Minstrel regarding TPC, this is
the best way to handle it, otherwise the least transmit power would be
used for every packet.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>