minstrel_ht-api: add manual tpc, rc-, tpc-only cmd
This patch introduces the manual mode also for TPC. It can be set via a
separate command for all or specific stations, is included in sta;add
and sta;dump lines and is independent of the RC manual mode, i.e. the
one can be set without the other, or both can be used at the same time.
The API-internal ratetable is used to store the TX power, thus some
modifications were required to the rate-setting process, e.g. not
freeing and reallocating the table after each mode change, and using
minstrel_ht_update_rates instead of rate_control_set_rates directly.

Two new commands 'set_rates' and 'set_power' are added to be able to set
TX rate or TX power without touching the other one. 'set_rates' is in
fact now identical in functionality to the original 'rates' command.
The control commands 'set_rates', 'set_power' and 'set_rates_power' fail
in case RC and TPC are not in manual mode accordingly.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>