minstrel_ht-api: use empty fields in txs lines
Previously, empty and unused mrr stages in txs lines were filled with
'ffff;0;ffff' as placeholder for these entries. However, this wastes
quite some space and is not necessary because the output is mostly
processed by parsers. They should be able to distinguish the different
fields without placeholders.
This should also have an effect on traffic amount caused by the API.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>