minstrel_ht-api: use , as stage separator in txs
Previously, the separator of the single values of an MRR stage was
equivalent to the overall used separator ';', e.g.
';rate0;count0;txpwr0;rate1;count1;txpwr1'.
To be consistent with the recent command format changes and to have a
better distinction, use ',' as the separator within a stage. Thus, the
mentioned example would look like
';rate0,count0,txpwr0;rate1,count1,txpwr1'.
At least in the kernel-side code this format is easier and less
error-prone to parse.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>