minstrel_ht-api: use empty field for invalid txpwr
Wireless drivers can report '-1' in case they are not able to report a
tx power for a tx status. This value resulted in 'ffff' as it was casted
to u16 before printing.
To be consistent with the output for unused MRR stages, no value will be
printed in that case anymore. E.g. ';97,2,ffff;' will now be ';97,2,;'.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>