orca: fix broken lines by adding missing lock
In the API output there were several broken lines where a
txs/rxs/stats line was being written but was interrupted and a
sample_rates line starts. This was caused by a missing lock in the
function that issues the sample_rates line to the relay so it could
occur while another issue to the relay is in progress.
In the place where this is called, e.g. the function for issuing
'best_rates' is also called before and it has a proper locking.

Adding the missing lock fixes the issue. Tested briefly that this
doesn't produce further deadlocks.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>