mac80211: reorder and renumber scnx patches
Patches are reordered so patches regarding the mac80211 layer itself
come first and after that patches regarding the API. Thus, most patches
regarding TPC are now coming before the ORCA patches. It is more likely
that the TPC patches may get upstream and they are less likely to change
that often.

Numbering of the patches is also changed, TPC and the API have now
higher numbers as they are completely new features.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>