orca_uapi: remove ff from rxs report
The rxs traces report several RSSI values and they are typically
negative. The s8 values are directly passed to printf and '%x' is used
as the format string. Because HEX output is used, printf casts the s8 to
an int and prints 4 bytes instead of one, thus the output always
contains 'ffffff' in front of the actual value. This is not necessary
and avoidable overhead.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>