orca_uapi: export stations of PHY via api_phy
Using ORCA UAPI, one can now get a list of stations with the same
information that is included in the 'sta;add' and 'sta;dump'. However,
one does not need to trigger them. Instead, they can be read from
api_phy with one line per STA.
The purpose of this is mainly to allow orca-rcd to retrieve the STA info
when a new client connects without having to trigger the aforementioned
line through api_control.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>