mt7615: add fine-grained TPC infrastructure
This adds the infrastructure to control the transmit power per packet
for mt7615-based WiFi chipsets.

Support for TPC is propagated to the mac80211 layer with the supported
TX power levels and code is added in the TX control path that sets the
TX power in the TXWI in case it was specified in the rate table or
tx_info. TPC is disabled by default, thus there should be no difference
in operation after this commit. It can be enabled via debugfs or by
setting the tpc_en flag in struct mt7615_phy.

mt7615 supports specifying a power offset between -8 dBm and +7.5 dBm
that is added to a rate-specific base power. Power index 0 corresponds
to -8 dBm and a range of power indices 0 - 31 can be used for up to
+7.5 dBm. By default (and when -1 is specified as TX power to be used),
the driver uses the default power index 16, which corresponds to a power
offset of 0 dBm.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>