orca_uapi: export controllable features via api_phy
This exports the current state of features that can be controlled
through the feature control interface through api_phy. A new line type
'ftrs' is added. The payload of this line begins with either 'on' or
'off', denoting the subtype of the line. Thus, lines starting with
'ftrs;on' provide a list of active features and lines starting with
'ftrs;off' provide a list of inactive features.
In case there are features that are supported by the mac80211 layer but
not by the hardware/driver, they do not occur in any of the lists and
thus are unsupported.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>