orca_uapi: check if TPC is ready before using it
Before, the commands for setting TX power and switching the API's TPC
mode didn't check whether TPC is even supported or, if driver supports
it, it is active.
This adds checks prior to setting TX power and switching TPC mode to
only continue command execution when TPC is supported and enabled.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>