orca_uapi: print sta-events regardless of monitoring
If a client hasn't yet enabled any monitoring mode and an STA is added,
updated or removed, the client may not recognize this because such
events are currently only printed if at least one monitoring mode is
active. This was originally supposed to be used with the MONITOR_STA
mode which was removed before.

This removes the checks for any active monitor mode before printing the
events, thus the events will be always printed from now on.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>