orca_uapi: make more independent from Minstrel
This makes the ORCA UAPI more independent from Minstrel as much as
possible to allow using a subset of ORCA's functionality with WiFi
chipsets that do not use Minstrel RC. This requires to place ORCA's
hooks in mac80211 instead of Minstrel and use alternative structures
or maintain Minstrel structures on our own when Minstrel doesn't run.

This commit also includes a patch for mt76 that fixes the issue that the
ieee80211_rate_status that is passed in mt76_tx_status_unlock has a
try_count of 0 and also a txpower_idx of 0. Setting these values to 1
and -1 respectively was missing. This patch, at least the change for the
try_count, should be proposed upstream.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>