orca_uapi: add patch version to API version info
A patch version is added to the existing line like 'orca_version;X;Y'
(where Y is the patch version). The patch version must be incremented
with every change to ensure the APIs current patch state running on
different nodes can be clearly distinguished. This is escpecially
helpful in our testing network to find out which nodes aren't running
the latest version.

While the major version (in the example above X) is only intended to be
increased for major breaking changes, the patch version must be used for
every change. In case the major version is increased, the patch version
should be reset to '0' or '1'.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>