mac80211: improve feature control interface
This again improves the feature control interface regarding several
aspects. Most notably it is now possible to set arbitrary values as
state instead of only 'on' and 'off'.

For ORCA UAPI, the values must be given in HEX and 'on' and 'off' are
discouraged from now on, being replaced by 1 and 0;

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>