orca_uapi: print current configured power limit in api_phy
This reads the current configured power limit via the get_txpower
hardware operation and prints a separate line in api_phy for this.
The power limit is given in 0.5 dBm steps. All drivers that implement
the get_txpower operation internally use this stepsize so the result is
always valid, and it can be easily used in particular for ath9k.

This commit also includes a patch for ath9k allowing to call get_txpower
without a reference to a VIF because it is not available where we call
it. Other drivers (e.g. mt76) do not even use that VIF reference so it
should be fine.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>