orca_uapi: ensure no valid power is passed in auto tpc_mode
The member txpower_idx in the rate table which is initialized by
Minstrel was not set to -1, instead it remains 0 because of kzalloc.
This can cause issues where in auto tpc_mode 0 is passed instead of -1.

This is fixed by properly initializing the ratetbl upon allocation and
by also ignoring existing values in case tpc_mode is set to auto.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>