orca_uapi: add command to get value or property
This adds a new command called 'get' that can be sent to api_control to
read/get any of the supported values. If the value could be read, the
API will reply with a global response line like
'<timestamp>;got;<property>;<value>', thus it will be received by
everyone who captures api_control output.

For the beginning, the single supported value/property is 'pwr-limit'
for reading the current maximum TX power limit. This value is equal to
the 'pwr-limit' given in api_phy but should not be confused with other
TX power values.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>