orca_uapi: add command for dumping features
This adds the command 'dump_ftrs' to dump the features with their
current states, similar to what is printed in api_phy. This can be
useful in combination with orca_rcd as client only get the initial
feature list once in the beginning but have no way to query it later.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>