minstrel_ht: fix OFDM rate mapping with rate_info
For looping over the ofdm_rates array, ARRAY_SIZE(mp->ofdm_rates) was
used. However, mp->ofdm_rates is a two-dimensional array with the bands
as the first dimension and the actual rates as the second dimension.
Thus, use ARRAY_SIZE(mp->ofdm_rates[0]) instead to loop over the rates
instead of the bands.

In addition, the index that is stored in mp->ofdm_rates, was incorrectly
used to get the bitrate from minstrel_ofdm_rates. This breaks if the
2.4GHz band is used as the indices are shifted to be usable with the
bitrates in the corresponding sband structure.

This fixes some issues where TX status reporting doesn't work properly
when using the newer ieee80211_rate_status reporting with OFDM rates.

This should be submitted upstream in the Linux kernel.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>